feat: add roi
