+++ packages/web/server/queries/osmosis/concentratedliquidity/index.ts
@@ -1,2 +1,3 @@
+export * from "./position-by-id";
 export * from "./positions-by-address";
 export * from "./unbonding-positions";
